Refactor: Player, UITest (DataProvider)
