+++ Assets/Scripts/Tests/Runtime/CH1/Player.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 312393aa941b4cca9a797b80e02ba188
+timeCreated: 1704906998
